build: port CTP2 to macOS Apple Silicon
- Replace Windows APIs with POSIX equivalents in debug tools
- Add COM compatibility layer (objbase.h) for non-Windows systems
- Fix GUID definition to be a proper 16-byte struct
- Fix template static member definitions for Clang
- Resolve duplicate symbol g_theWonderMovieDB
- Fix MapPointData layout to include z coordinate consistently
- Add missing gfx/spritesys files to meson.build
- Add -D_DEBUG compiler flag for debug builds
- Switch build system to SDL2 and SDL2_mixer on macOS
- Add Homebrew dependencies (Brewfile)
- Many compilation fixes for macOS/ARM64 (case-sensitive includes,
  missing headers, implicit int, format strings, etc.)
- Executable links successfully: 20MB Mach-O ARM64

+++ ctp2_code/ctp/debugtools/debugassert.h
@@ -16,6 +16,7 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
 #include "ctp/debugtools/log.h"
 
 
@@ -69,8 +70,8 @@ void DebugAssert_Close (void);
 
 
 #undef ASSERT
-#define ASSERT(condition)													\
-	{if (! ((int) (condition))) {											\
+#define ASSERT(condition)											\
+	{if (! ((intptr_t) (condition))) {										\
 		DebugAssert_Assert ( __FILE__, __LINE__, LOG_ASSERT, #condition);	\
 	}}
 
@@ -90,7 +91,7 @@ void DebugAssert_Close (void);
 
 
 #define ASSERT_CLASS(log_class, condition)								\
-	{if (! ((int) (condition))) {										\
+	{if (! ((intptr_t) (condition))	) {									\
 		DebugAssert_Assert (__FILE__, __LINE__, log_class, #condition);	\
 	}}
 
@@ -107,7 +108,7 @@ void DebugAssert_Close (void);
 
 
 #define ASSERT_INDIRECT(module_name, module_line, condition)					\
-	{if (! ((int) (condition))) {												\
+	{if (! ((intptr_t) (condition))	) {											\
 		DebugAssert_Assert (module_name, module_line, LOG_ASSERT, #condition);	\
 	}}
